server/src/requirements.txt
pyramid==1.10.5
pyramid-jinja2==2.8
opencv-python-headless
mysql-connector-python